Fixed up landing page for demo, added FieldPopulatorSupport
